+++ src/OrbitGl/TriangleToggle.cpp
@@ -5,8 +5,8 @@
 #include "TriangleToggle.h"
 
 #include <GteVector.h>
-#include <math.h>
 
+#include <cmath>
 #include <utility>
 
 #include "AccessibleTriangleToggle.h"
@@ -36,10 +36,10 @@ void TriangleToggle::DoDraw(PrimitiveAssembler& primitive_assembler, TextRendere
   // Draw triangle.
   const Vec2 midpoint = GetPos() + Vec2(GetWidth() / 2, GetHeight() / 2);
   const float triangle_side_length = std::min(GetWidth(), GetHeight());
-  const float kCos30 = sqrtf(3.f) / 2;
+  const float kCos30 = std::sqrt(3.f) / 2;
   const float triangle_height = triangle_side_length * kCos30;
 
-  const float half_triangle_height = ceilf(triangle_height / 2);
+  const float half_triangle_height = std::ceil(triangle_height / 2);
   const float half_triangle_side_length = triangle_side_length / 2;
 
   if (!picking) {
